requirements.txt
streamlit
python-dotenv
requests
pillow
fpdf